+++ src/lib/output/plugins/JavascriptIndexPlugin.ts
@@ -2,15 +2,27 @@ import * as Path from "path";
 import { Builder, trimmer } from "lunr";
 
 import {
+    Comment,
     DeclarationReflection,
     ProjectReflection,
     ReflectionKind,
 } from "../../models";
 import { GroupPlugin } from "../../converter/plugins";
 import { Component, RendererComponent } from "../components";
-import { RendererEvent } from "../events";
-import { writeFileSync } from "../../utils";
+import { IndexEvent, RendererEvent } from "../events";
+import { BindOption, writeFileSync } from "../../utils";
 import { DefaultTheme } from "../themes/default/DefaultTheme";
+
+/**
+ * Keep this in sync with the interface in src/lib/output/themes/default/assets/typedoc/components/Search.ts
+ */
+interface SearchDocument {
+    kind: number;
+    name: string;
+    url: string;
+    classes?: string;
+    parent?: string;
+}
 
 /**
  * A plugin that exports an index of the project to a javascript file.
@@ -19,6 +31,9 @@ import { DefaultTheme } from "../themes/default/DefaultTheme";
  */
 @Component({ name: "javascript-index" })
 export class JavascriptIndexPlugin extends RendererComponent {
+    @BindOption("searchInComments")
+    searchComments!: boolean;
+
     /**
      * Create a new JavascriptIndexPlugin instance.
      */
@@ -39,30 +54,52 @@ export class JavascriptIndexPlugin extends RendererComponent {
             return;
         }
 
-        const rows: any[] = [];
+        const rows: SearchDocument[] = [];
         const kinds: { [K in ReflectionKind]?: string } = {};
 
-        for (const reflection of event.project.getReflectionsByKind(
-            ReflectionKind.All
+        const initialSearchResults = Object.values(
+            event.project.reflections
+        ).filter((refl) => {
+            return (
+                refl instanceof DeclarationReflection &&
+                refl.url &&
+                refl.name &&
+                !refl.flags.isExternal
+            );
+        }) as DeclarationReflection[];
+
+        const indexEvent = new IndexEvent(
+            IndexEvent.PREPARE_INDEX,
+            initialSearchResults
+        );
+
+        this.owner.trigger(indexEvent);
+
+        if (indexEvent.isDefaultPrevented) {
+            return;
+        }
+
+        const builder = new Builder();
+        builder.pipeline.add(trimmer);
+
+        builder.ref("id");
+        for (const [key, boost] of Object.entries(
+            indexEvent.searchFieldWeights
         )) {
-            if (!(reflection instanceof DeclarationReflection)) {
+            builder.field(key, { boost });
+        }
+
+        for (const reflection of indexEvent.searchResults) {
+            if (!reflection.url) {
                 continue;
             }
 
-            if (
-                !reflection.url ||
-                !reflection.name ||
-                reflection.flags.isExternal
-            ) {
-                continue;
-            }
-
-            let parent = reflection.parent;
             const boost = reflection.relevanceBoost ?? 1;
             if (boost <= 0) {
                 continue;
             }
 
+            let parent = reflection.parent;
             if (parent instanceof ProjectReflection) {
                 parent = undefined;
             }
@@ -73,33 +110,28 @@ export class JavascriptIndexPlugin extends RendererComponent {
                 );
             }
 
-            const row: any = {
-                id: rows.length,
+            const row: SearchDocument = {
                 kind: reflection.kind,
                 name: reflection.name,
                 url: reflection.url,
                 classes: reflection.cssClasses,
             };
 
-            if (boost !== 1) {
-                row.boost = boost;
-            }
-
             if (parent) {
                 row.parent = parent.getFullName();
             }
 
+            builder.add(
+                {
+                    name: reflection.name,
+                    comment: this.getCommentSearchText(reflection),
+                    ...indexEvent.searchFields[rows.length],
+                    id: rows.length,
+                },
+                { boost }
+            );
             rows.push(row);
         }
-
-        const builder = new Builder();
-        builder.pipeline.add(trimmer);
-
-        builder.ref("id");
-        builder.field("name", { boost: 10 });
-        builder.field("parent");
-
-        rows.forEach((row) => builder.add(row));
 
         const index = builder.build();
 
@@ -119,5 +151,30 @@ export class JavascriptIndexPlugin extends RendererComponent {
             jsonFileName,
             `window.searchData = JSON.parse(${JSON.stringify(jsonData)});`
         );
+    }
+
+    private getCommentSearchText(reflection: DeclarationReflection) {
+        if (!this.searchComments) return;
+
+        const comments: Comment[] = [];
+        if (reflection.comment) comments.push(reflection.comment);
+        reflection.signatures?.forEach(
+            (s) => s.comment && comments.push(s.comment)
+        );
+        reflection.getSignature?.comment &&
+            comments.push(reflection.getSignature.comment);
+        reflection.setSignature?.comment &&
+            comments.push(reflection.setSignature.comment);
+
+        if (!comments.length) {
+            return;
+        }
+
+        return comments
+            .flatMap((c) => {
+                return [...c.summary, ...c.blockTags.flatMap((t) => t.content)];
+            })
+            .map((part) => part.text)
+            .join("\n");
     }
 }